+++ src/main/java/com/itca/cursify/controller/UserController.java
@@ -1,27 +1,26 @@
 package com.itca.cursify.controller;
 
-import com.itca.cursify.persistece.entity.Role;
 import com.itca.cursify.persistece.entity.User;
-import com.itca.cursify.service.RoleService;
-import com.itca.cursify.service.UserServices;
-import com.itca.cursify.service.dto.RoleInDTO;
+import com.itca.cursify.service.UserService;
 import com.itca.cursify.service.dto.UserInDTO;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/v1/users")
 public class UserController {
-    private final UserServices userServices;
+    private final UserService userService;
 
-    public UserController(UserServices userServices) {
-        this.userServices = userServices;
+    public UserController(UserService userService) {
+        this.userService = userService;
     }
+
     @PostMapping("/create")
     public User createUser(@RequestBody UserInDTO userInDTO) {
-        return this.userServices.createUser(userInDTO);
+        return this.userService.createUser(userInDTO);
     }
 
+    @PostMapping("/login")
+    public User login(@RequestParam String email, @RequestParam String password) {
+        return userService.findUserByEmailAndPassword(email, password);
+    }
 }
